# crates/tellurion-server/openapi_stac.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tellurion — STAC API (core + collections + items + search + assets)",
    "description": "STAC API - Core (landing page, conformance, OpenAPI) plus STAC API - Collections, STAC API - Features (items), STAC API - Item Search (/search, both GET and POST, with the Filter Extension and free-text query), and this server's assets-and-object-storage surface (registration, direct upload, presigned upload, resumable upload, download redirect, and the read-only reconcile report) for one catalog. Mounted at /{tenant}/stac/catalogs/{catalog}.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/collections/{collectionId}/assets/{key}": {
      "get": {
        "summary": "Fetch one collection asset's registration",
        "description": "Collection-level reads check configuration-declared assets first: a declared asset is a remote asset that happens to live in configuration — same read surface, always available, no byte lifecycle.",
        "operationId": "getCollectionAsset",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          },
          {
            "$ref": "#/components/parameters/assetKey"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/assetObject"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          }
        }
      },
      "put": {
        "summary": "Register (or re-register) one collection asset",
        "description": "Managed vs. remote is discriminated by the Repr-Digest header, not by a body field. A managed registration is born pending and carries no bytes yet; the transfer happens on one of the .../data routes below. A remote registration is born available and has no byte lifecycle.",
        "operationId": "putCollectionAsset",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          },
          {
            "$ref": "#/components/parameters/assetKey"
          },
          {
            "$ref": "#/components/parameters/reprDigest"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/assetObjectInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/assetObject"
          },
          "400": {
            "$ref": "#/components/responses/badRequest"
          },
          "404": {
            "$ref": "#/components/responses/capabilityUnsupported"
          },
          "409": {
            "$ref": "#/components/responses/conflict"
          },
          "413": {
            "$ref": "#/components/responses/payloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/unsupportedMediaType"
          }
        }
      },
      "delete": {
        "summary": "Remove one collection asset",
        "description": "A remote asset loses its record only; a managed asset loses its record and its stored object.",
        "operationId": "deleteCollectionAsset",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          },
          {
            "$ref": "#/components/parameters/assetKey"
          }
        ],
        "responses": {
          "204": {
            "description": "The asset registration was removed."
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "409": {
            "$ref": "#/components/responses/conflict"
          }
        }
      }
    },
...
    "/collections/{collectionId}/items/{itemId}/assets/{key}": {
      "get": {
        "summary": "Fetch one item asset's registration",
        "description": "Reads the item-level asset record; configuration-declared assets are a collection-level concept and never apply here.",
        "operationId": "getItemAsset",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          },
          {
            "$ref": "#/components/parameters/itemId"
          },
          {
            "$ref": "#/components/parameters/assetKey"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/assetObject"
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          }
        }
      },
      "put": {
        "summary": "Register (or re-register) one item asset",
        "description": "Managed vs. remote is discriminated by the Repr-Digest header, not by a body field. A managed registration is born pending and carries no bytes yet; the transfer happens on one of the .../data routes below. A remote registration is born available and has no byte lifecycle.",
        "operationId": "putItemAsset",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          },
          {
            "$ref": "#/components/parameters/itemId"
          },
          {
            "$ref": "#/components/parameters/assetKey"
          },
          {
            "$ref": "#/components/parameters/reprDigest"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/assetObjectInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/assetObject"
          },
          "400": {
            "$ref": "#/components/responses/badRequest"
          },
          "404": {
            "$ref": "#/components/responses/capabilityUnsupported"
          },
          "409": {
            "$ref": "#/components/responses/conflict"
          },
          "413": {
            "$ref": "#/components/responses/payloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/unsupportedMediaType"
          }
        }
      },
      "delete": {
        "summary": "Remove one item asset",
        "description": "A remote asset loses its record only; a managed asset loses its record and its stored object.",
        "operationId": "deleteItemAsset",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          },
          {
            "$ref": "#/components/parameters/itemId"
          },
          {
            "$ref": "#/components/parameters/assetKey"
          }
        ],
        "responses": {
          "204": {
            "description": "The asset registration was removed."
          },
          "404": {
            "$ref": "#/components/responses/notFound"
          },
          "409": {
            "$ref": "#/components/responses/conflict"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "collectionId": {
        "name": "collectionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "itemId": {
        "name": "itemId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "assetKey": {
        "name": "key",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "The asset key within this collection (or item). A key declared in configuration is served here read-only: PUT and DELETE against it refuse with a named 409."
      },
      "reprDigest": {
        "name": "Repr-Digest",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "RFC 9530 representation digest. Its presence is what discriminates a managed registration (digest present: no `href`, `file:size` required, this server will hold the bytes) from a remote one (digest absent: `href` required, born available, no byte lifecycle)."
      },
...
    },
    "responses": {
      "badRequest": {
        "description": "Invalid request parameter.",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
...
      "notFound": {
        "description": "Unknown collection or item, or the driver doesn't support any capability this route needs.",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "assetObject": {
        "description": "The asset's current registration, including its lifecycle status.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/assetObject"
            }
          }
        }
      },
...
      "conflict": {
        "description": "The key is a configuration-declared asset (read-only through this API), or the request conflicts with the asset's current state.",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "payloadTooLarge": {
        "description": "The request body exceeds the applicable cap: this collection's request-body budget on a registration, or the asset's own declared size (minus what is already stored) on a transfer.",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "unsupportedMediaType": {
        "description": "The declared media type is not in this collection's asset media-type allow-list.",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object"
            }
          }
        }
      },
      "capabilityUnsupported": {
        "description": "This collection has no asset store, no object store, or an object store whose profile lacks the capability this route needs (presigned URLs, resumable uploads, or listing) — refused by name, never as a 500.",
        "content": {
          "application/problem+json": {
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "schemas": {
      "assetObjectInput": {
        "type": "object",
        "description": "A STAC Asset Object plus the STAC `file` extension's `file:size`.",
        "properties": {
          "href": {
            "type": "string",
            "description": "Required for a remote registration; MUST be absent for a managed one — the server derives the eventual .../data href itself."
          },
          "type": {
            "type": "string",
            "description": "Media type, checked against this collection's asset media-type allow-list on a managed registration (415 when refused)."
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "file:size": {
            "type": "integer",
            "minimum": 0,
            "description": "Declared byte length, required for a managed registration and checked against this collection's asset size cap (413 when refused). It is the cap every later transfer on this key is measured against."
          }
        }
      },
      "assetObject": {
        "type": "object",
        "description": "A STAC Asset Object plus this extension's own lifecycle `status`/`status_detail`.",
        "required": [
          "href",
          "status"
        ],
        "properties": {
          "href": {
            "type": "string",
            "description": "For a managed asset, this server's own .../data URL; for a remote one, the registered external URL."
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "available",
              "failed"
            ]
          },
          "status_detail": {
            "type": "string"
          }
        }
      },
...
    }
  }
}